+++ src/Contract/SettingRepository.php
@@ -18,6 +18,8 @@ declare(strict_types=1);
 
 namespace SettingsManager\Contract;
 
+use SettingsManager\Exception\SettingValueNotFoundException;
+
 /**
  * Setting repository interface defines the interface for storing raw setting values in a persistence layer
  *
@@ -32,6 +34,15 @@ interface SettingRepository
      * @return mixed|null
      */
     public function findValue(string $settingName);
+
+    /**
+     * Get a setting value by its name or throw an exception if not found
+     *
+     * @param string $settingName
+     * @return mixed
+     * @throws SettingValueNotFoundException
+     */
+    public function getValue(string $settingName);
 
     /**
      * List values