+++ src/Registry/SettingDefinitionRegistry.php
@@ -22,7 +22,7 @@ use ArrayIterator;
 use Countable;
 use IteratorAggregate;
 use SettingsManager\Contract\SettingDefinition;
-use SettingsManager\Exception\SettingDefinitionNotFoundException;
+use SettingsManager\Exception\UndefinedSettingException;
 use SettingsManager\Exception\SettingNameCollisionException;
 
 /**
@@ -66,7 +66,7 @@ class SettingDefinitionRegistry implements IteratorAggregate, Countable
         if ($this->has($name)) {
             return $this->items[$name];
         } else {
-            throw SettingDefinitionNotFoundException::forSetting($name);
+            throw UndefinedSettingException::forSetting($name);
         }
     }